// src/app/details/bench.interface.ts
export interface Bench {
    lab_obj_id:string;
    bench_id: string;
    poc: string;
    technology: string;
    type: string;
    cell_id:string;
    rru_output_db: string;
    attenuation_db:string;
    real_time_signal: string;
    comments: string;
    utilization: string;
    enabled_date: string;
}